Configuration package renamed to config
